requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
astroid==3.3.9
attrs==25.3.0
beautifulsoup4==4.12.2
billiard==4.2.1
black==23.11.0
celery==5.3.4
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.8.0
dill==0.3.9
Django==5.2
django-celery-results==2.5.1
djangorestframework==3.15.0
djangorestframework_simplejwt==5.5.0
h11==0.14.0
idna==3.10
iniconfig==2.1.0
isort==5.12.0
kombu==5.5.2
lxml==4.9.3
mccabe==0.7.0
mypy==1.7.1
mypy-extensions==1.0.0
outcome==1.3.0.post0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
pluggy==1.5.0
prompt_toolkit==3.0.50
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.9.0
pylint==3.3.6
PySocks==1.7.1
pytest==7.4.3
pytest-cov==4.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
redis==5.0.0
requests==2.31.0
selenium==4.15.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
tomlkit==0.13.2
trio==0.29.0
trio-websocket==0.12.2
types-beautifulsoup4==4.12.0.0
types-html5lib==1.1.11.20241018
types-requests==2.31.0.2
types-urllib3==1.26.25.14
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
wsproto==1.2.0
